move SvgDisplayInteractivity into SvgDisplay tests
